+++ src/main.ts
@@ -18,7 +18,7 @@ import { HeadlessScreenshotHelperOptions, STYLIZED_POSTPROCESSING, defaultCanvas
 import { setFSModule } from 'molstar/lib/commonjs/mol-util/data-source';
 
 import { PDBeAPI } from './api';
-import { Args, ImageTypes, Mode, Modes } from './args';
+import { Args, Defaults, ImageTypes, Mode, Modes, createArgs } from './args';
 import { collectCaptions } from './captions/collect';
 import { checkMissingFiles, getExpectedFiles } from './expected-files';
 import { fetchUrl, gunzipData, parseIntStrict } from './helpers/helpers';
@@ -37,8 +37,7 @@ const DEFAULT_INPUT_URL_TEMPLATES: { [mode in Mode]: string } = {
     pdb: 'https://www.ebi.ac.uk/pdbe/entry-files/download/${id}.bcif',
     alphafold: 'https://alphafold.ebi.ac.uk/files/${id}.cif', // There is some issue with AlphaFold bcifs, this might be fixed in the future
 };
-const DEFAULT_PDBE_API_URL = 'https://www.ebi.ac.uk/pdbe/api';
-const DEFAULT_IMAGE_SIZE = '800x800';
+const DEFAULT_IMAGE_SIZE_STRINGS = Defaults.size.map(s => `${s.width}x${s.height}`);
 
 
 /** Return parsed command line arguments for `main` */
@@ -49,14 +48,14 @@ export function parseArguments(): Args {
     parser.add_argument('output_dir', { help: 'Output directory.' });
     parser.add_argument('--input', { help: 'Input structure file path or URL (.cif, .bcif, .cif.gz, .bcif.gz).' });
     parser.add_argument('--input-public', { help: 'Input structure URL to use in saved Mol* states (.molj files) (cif or bcif format).' });
-    parser.add_argument('--mode', { choices: [...Modes], default: 'pdb', help: 'Mode.' });
-    parser.add_argument('--api-url', { default: DEFAULT_PDBE_API_URL, help: `PDBe API URL. Default: ${DEFAULT_PDBE_API_URL}.` });
+    parser.add_argument('--mode', { choices: [...Modes], default: Defaults.mode, help: 'Mode.' });
+    parser.add_argument('--api-url', { default: Defaults.api_url, help: `PDBe API URL. Default: ${Defaults.api_url}.` });
     parser.add_argument('--api-retry', { action: 'store_true', help: 'Retry any failed API call up to 5 times, waiting random time (up to 30 seconds) before each retry.' });
     parser.add_argument('--no-api', { action: 'store_true', help: 'Do not use PDBe API at all (some images will be skipped, some entity names will be different in captions, etc.).' });
-    parser.add_argument('--size', { nargs: '*', default: [DEFAULT_IMAGE_SIZE], help: `One or more output image sizes, e.g. 800x800 200x200. Default: ${DEFAULT_IMAGE_SIZE}. Only the largest size is rendered, others are obtained by resizing unless --render_each_size is used. Use without any value to disable image rendering (only create captions and MOLJ files).` });
+    parser.add_argument('--size', { nargs: '*', default: DEFAULT_IMAGE_SIZE_STRINGS, help: `One or more output image sizes, e.g. 800x800 200x200. Default: ${DEFAULT_IMAGE_SIZE_STRINGS.join(' ')}. Only the largest size is rendered, others are obtained by resizing unless --render_each_size is used. Use without any value to disable image rendering (only create captions and MOLJ files).` });
     parser.add_argument('--render-each-size', { action: 'store_true', help: 'Render image for each size listed in --size, instead of rendering only the first size and resampling to the other sizes.' });
-    parser.add_argument('--type', { nargs: '*', choices: [...ImageTypes], default: ['all'], help: 'One or more image types to be created. Use "all" as a shortcut for all types. See README.md for details on image types. Default: all. Use without any value to skip all types (only create summary files from existing outputs).' });
-    parser.add_argument('--view', { choices: ['front', 'all', 'auto'], default: 'auto', help: 'Select which views should be created for each image type (front view / all views (front, side, top) / auto (creates all views only for these image types: entry, assembly, entity, modres, plddt)). Default: auto.' });
+    parser.add_argument('--type', { nargs: '*', choices: [...ImageTypes], default: Defaults.type, help: `One or more image types to be created. Use "all" as a shortcut for all types. See README.md for details on image types. Default: ${Defaults.type.join(' ')}. Use without any value to skip all types (only create summary files from existing outputs).` });
+    parser.add_argument('--view', { choices: ['front', 'all', 'auto'], default: Defaults.view, help: `Select which views should be created for each image type (front view / all views (front, side, top) / auto (creates all views only for these image types: entry, assembly, entity, modres, plddt)). Default: ${Defaults.view}.` });
     parser.add_argument('--opaque-background', { action: 'store_true', help: 'Render opaque background in images (default: transparent background).' });
     parser.add_argument('--no-axes', { action: 'store_true', help: 'Do not render axis indicators aka PCA arrows (default: render axes when rendering the same scene from multiple view angles (front, side, top)).' });
     parser.add_argument('--show-hydrogens', { action: 'store_true', help: 'Show hydrogen atoms in ball-and-stick visuals (default: always ignore hydrogen atoms).' });
@@ -65,7 +64,7 @@ export function parseArguments(): Args {
     parser.add_argument('--allow-lowest-quality', { action: 'store_true', help: "Allow any quality level for visuals, including 'lowest', which is really ugly (default: allow only 'lower' quality level and better)." });
     parser.add_argument('--date', { help: `Date to use as "last_modification" in the caption JSON (default: today's date formatted as YYYY-MM-DD).` });
     parser.add_argument('--clear', { action: 'store_true', help: 'Remove all contents of the output directory before running.' });
-    parser.add_argument('--log', { choices: [...LogLevels], type: (s: string) => s.toUpperCase(), default: 'INFO', help: 'Set logging level. Default: INFO.' });
+    parser.add_argument('--log', { choices: [...LogLevels], type: (s: string) => s.toUpperCase(), default: Defaults.log, help: `Set logging level. Default: ${Defaults.log}.` });
     const args = parser.parse_args();
     args.size = args.size.map((s: string) => {
         try {